requirements.txt
streamlit
tweepy
pillow